[bugfix] LTX-2: restore input dtype in _rms_norm_dispatch under torch >= 2.12
torch 2.12 moved rms_norm to autocast's float32 cast policy, so the
block-level functional norm started returning float32 under
autocast(bfloat16). The NVFP4 prequant path feeds this output straight
into quantize_input (bypassing autocast's recast on the next linear),
which asserts "only allow bf16/fp16 inputs to fp4 linear".
StageAwareRMSNorm already carried this fix; apply the same dtype
restore to _rms_norm_dispatch. No-op on torch <= 2.11.

+++ fastvideo/models/dits/ltx2.py
@@ -72,8 +72,14 @@ def _rms_norm_dispatch(
     """Use QuACK RMSNorm only for LTX-2 refine stage, else Torch RMSNorm."""
     if _is_ltx2_refine_stage():
         return _quack_rmsnorm(x, weight=weight, eps=eps)
+    # torch 2.12 added rms_norm to autocast's float32 cast policy, so under
+    # autocast(bfloat16) this returns float32 where torch <= 2.11 returned
+    # bfloat16. Restore the input dtype so autocast-blind consumers (the
+    # NVFP4 quantize_input prequant path feeds this straight into the FP4
+    # kernel) don't receive upcast activations. Same fix as
+    # StageAwareRMSNorm.forward below.
     return torch.nn.functional.rms_norm(
-        x, (x.shape[-1], ), weight=weight, eps=eps)
+        x, (x.shape[-1], ), weight=weight, eps=eps).to(x.dtype)
 
 
 class StageAwareRMSNorm(nn.RMSNorm):
